feat: Add EncryptionManager utility class for basic encryption-decryption workflow
When encrypting a file, the input file is accessed, encrypted in-place and the encryption key used (randomly generated) is returned.

When decrypting a file, the input file and the encryption key are passed in. The input file's content is read, and then the decrypted content is saved to a 'decrypted/' folder in the media/ directory. A url to the decrypted file is returned.

In the future, the decrypted content would not be saved. This is just for testing the workflow.